pyproject.toml
[tool.poetry]
name = "vt-grass-bot"
version = "0.1.0"
description = ""
authors = ["SSSKrut"]
license = "WTFPL"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.10.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"